Added custom hook resize observer
