+++ react-app-part-one/src/App.js
@@ -1,96 +1,22 @@
-import React, { useRef, useEffect, useState} from 'react';
-import {select,  axisBottom, axisRight, scaleLinear, scaleBand} from 'd3';
+import React, { useState} from 'react';
+import BarChart from "./Barchart";
 import './App.css';
 
 
 
 function App() {
 
-  const svgHeight = 200;
-  const svgWidth = 400;
-  const paddingLeft = 10;
-  const paddingRight = 30;
-  const paddingBottom = 20;
   const initialData = [35,5,15,60,20,40,10,75,60,32];
   const randomData = [...Array(10)].map(element=>~~(Math.random()*125)); //
   const [dataSet, setData] = useState(initialData);
-  const svgRef = useRef();
-
-
-  useEffect( () => {
-    const maxValue = Math.max(...initialData);
-    const highestYValue = svgHeight - maxValue+paddingBottom;
-    const svg = select(svgRef.current);
-    // Scale band takes value specified in   
-    // map arbitrary value to a range of linear values
-    //Need to be explicit about what to map ex domain([0,1,2,3])
-    // divide the values into equal bands
-    const xScale = scaleBand()
-      .domain(dataSet.map((element, index) => index))
-      .range([paddingLeft, svgWidth-paddingRight])
-      .padding(0.5); // Scalband() takes a padding to separate bands
-    
-    /**
-     * adding ColorScale
-     */
-    const colorScale = scaleLinear()
-                    .domain([30,~~(highestYValue/2),highestYValue])
-                    .range(["#C5EDAC","#F7A278","orange"])
-                    .clamp(true); // forces values defined in the domain to remain thecolors
-
-
-    const yScale = scaleLinear()
-                    .domain([0,highestYValue])
-                    .range([svgHeight-paddingBottom, 0]);  
-
-   
-    const xAxis = axisBottom(xScale).ticks(dataSet.length);
-    svg.select(".x-axis")
-        .style('transform',` translateY(${svgHeight-paddingBottom}px)` )
-        .call(xAxis);
-
-   const yAxis = axisRight(yScale);
-   svg.select(".y-axis")
-            .style('transform',` translateX(${svgWidth-paddingRight}px)` )
-            .call(yAxis);
-
-  // Draw Bar
-  svg.selectAll(".bar")
-      .data(dataSet)
-        .join('rect')
-          .attr('class','bar')
-          .attr('transform','scale(1, -1)')//flip the bar upside down to fix wron animation start
-          .attr('x', (value,index) => xScale(index))
-          .attr('y', -svgHeight+paddingBottom)
-          .attr('width', xScale.bandwidth()) // bandwidth equals to the width of one band
-          .on("mouseenter", (value, index) => {
-             console.log(value);
-             svg.selectAll(".tooltip")
-                .data([value]) // Sync with one data element
-                .join( enter => enter.append("text").attr("y", yScale(value))) // .join("text") Creates a new text element, Adds animation by fixing Yscale of element entering
-                  .attr("class","tooltip") 
-                  .text(value) // Updates value of text element with current value of rect being hovered
-                  .attr("x",xScale(index) + xScale.bandwidth()/2) // display element on x scale, using index of the current element array
-                  .attr("text-anchor",'middle') // To center the tooltip in the center add  + xScale.bandwidth()/2
-                  .transition()
-                    .attr("y",yScale(value)-paddingLeft) // tooltip position on y axis
-                    .attr("opacity",1)
-                ;
-            })
-          .on("mouseleave", () => svg.select(".tooltip").remove())  
-          .transition()//transition will animate attribute called after it
-          .attr('fill',colorScale)  
-          .attr('height',value => svgHeight - yScale(value) - paddingBottom);  
-
-  },[dataSet,initialData,randomData]);
-
+  const data ={
+    dataSet: dataSet,
+    initialData :initialData,
+    randomData:randomData
+  }
   return (<React.Fragment>
-        <h2>Curved Line Chart: X and Y Axis</h2>
-        <svg width={svgWidth} height={svgHeight} ref={svgRef}>
-          <g className="x-axis"/>
-          <g className="y-axis"/>
-        </svg>
-        <br/>
+        <BarChart data={data}/>
+      
         <div className="button-container">
             <button onClick={ () => setData(dataSet.map( value => value+ 5))} >Update + 5</button>
             <span>&nbsp;</span>
